requirements.txt
cssutils==2.6.0
pandas==1.4.2
Pillow==9.5.0
qrcode==7.4.2
selenium==4.8.3
